[Chore] rename file to `editorGroupView.ts`
