+++ src/workbench/parts/workspace/editor/editorGroupView.ts
